server-side/requirements.txt
ultralytics==8.3.0
gradio>=4.0.0
onnxruntime>=1.18.0
opencv-python
numpy
pillow
pandas
matplotlib